* I can't believe it's so difficult to believe "I don't care if this fails"

+++ source/newio.c
@@ -751,7 +751,7 @@ static int	unix_recv (int channel, int quiet)
 	{
 		if (!quiet)
 		   syserr(CSRV(channel), "unix_recv: dgets_buffer(%d, %*s) failed",
-				channel, c, buffer);
+				channel, (int)c, buffer);
 		return -1;
 	}
